Add initial version of nodemoon file-watching utility
This commit introduces nodemoon, a simple file-watching utility similar to nodemon, intended for learning purposes. The main executable is index.js, which watches JavaScript files for changes and restarts a child process to execute them. The package.json and .gitignore files are also initialized to configure the project and exclude `.idea` directories, respectively.